service/pyproject.toml
[tool.poetry]
name = "docai"
version = "0.1.0"
description = "GCP's Doc AI wrapper written in Python based RESTful service using FastAPI"
authors = ["Nikhil Akki <nikhil.akki@outlook.com>"]
license = "MIT"

[tool.poetry.dependencies]
python = "^3.9"
fastapi = "^0.68.1"
uvicorn = "^0.15.0"
SQLAlchemy = "^1.4.25"
pydantic = {extras = ["email"], version = "^1.8.2"}
psycopg2-binary = "^2.9.1"
python-jose = "^3.3.0"
passlib = {extras = ["bcrypt"], version = "^1.7.2"}
python-dotenv = "^0.19.0"
python-multipart = "^0.0.5"
tenacity = "^8.0.1"
alembic = "^1.7.3"

[tool.poetry.dev-dependencies]
pylama = "^7.7.1"
black = "^21.9b0"
bandit = "^1.7.0"
pytest = "^6.2.5"
pytest-cov = "^2.12.1"
requests = "^2.26.0"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
